Move inlining
