+++ csrc/compute_at.h
@@ -7,15 +7,7 @@
 // clang-format on
 #pragma once
 
-#include <exceptions.h>
-#include <inlining.h>
-#include <logical_domain_map.h>
-#include <transform_replay.h>
-
-#include <deque>
-#include <unordered_map>
-#include <unordered_set>
-#include <vector>
+#include <ir/interface_nodes.h>
 
 namespace nvfuser {
