+++ csrc/scheduler/matmul.cpp
@@ -7,7 +7,6 @@
 // clang-format on
 #include <abstract_tensor.h>
 #include <device_lower/analysis/circular_buffer.h>
-#include <inlining.h>
 #include <instrumentation.h>
 #include <multidevice/utils.h>
 #include <scheduler/debug_utils.h>
@@ -15,6 +14,7 @@
 #include <scheduler/matmul_utils.h>
 #include <scheduler/mma_utils.h>
 #include <scheduler/multi_matmul.h>
+#include <scheduler/tools/inlining.h>
 #include <scheduler/utils.h>
 
 // NOTE: included to avoid compilation error caused by missing destructor in
