+++ csrc/scheduler/reduction_utils.cpp
@@ -8,13 +8,13 @@
 #include <scheduler/reduction_utils.h>
 
 #include <expr_evaluator.h>
-#include <inlining.h>
 #include <ir/cloner.h>
 #include <ir/utils.h>
 #include <maxinfo_propagator.h>
 #include <multidevice/utils.h>
 #include <ops/arith.h>
 #include <scheduler/registry.h>
+#include <scheduler/tools/inlining.h>
 #include <scheduler/utils.h>
 #include <transform_replay.h>
