+++ tests/cpp/test_circular_buffering.cpp
@@ -6,8 +6,8 @@
  */
 // clang-format on
 
-#include <inlining.h>
 #include <ops/all_ops.h>
+#include <scheduler/tools/inlining.h>
 #include <string.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
