+++ tests/cpp/test_combined_inner_outer_reduction.cpp
@@ -10,13 +10,13 @@
 #include <gtest/gtest.h>
 
 #include <grouped_reduction.h>
-#include <inlining.h>
 #include <ir/utils.h>
 #include <ops/all_ops.h>
 #include <runtime/fusion_executor_cache.h>
 #include <scheduler/all_schedulers.h>
 #include <scheduler/normalization_utils.h>
 #include <scheduler/reduction_utils.h>
+#include <scheduler/tools/inlining.h>
 #include <scheduler/utils.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
