+++ tests/cpp/test_fusion_profiler.cpp
@@ -11,11 +11,11 @@
 
 #include <fusion.h>
 #include <fusion_profiler.h>
-#include <inlining.h>
 #include <ops/all_ops.h>
 #include <runtime/executor.h>
 #include <runtime/executor_utils.h>
 #include <runtime/fusion_executor_cache.h>
+#include <scheduler/tools/inlining.h>
 #include <sys_utils.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
