+++ tests/cpp/test_gpu_fused_reduction.cpp
@@ -15,7 +15,6 @@
 #include <fusion.h>
 #include <fusion_segmenter.h>
 #include <grouped_reduction.h>
-#include <inlining.h>
 #include <ir/all_nodes.h>
 #include <ir/builder.h>
 #include <ir/graphviz.h>
@@ -30,6 +29,7 @@
 #include <runtime/fusion_executor_cache.h>
 #include <scheduler/all_schedulers.h>
 #include <scheduler/reduction_utils.h>
+#include <scheduler/tools/inlining.h>
 #include <scheduler/utils.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
