+++ tests/cpp/test_gpu_transpose.cpp
@@ -9,7 +9,6 @@
 #include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 
-#include <inlining.h>
 #include <ops/all_ops.h>
 #include <preseg_passes/allocation_order_inference.h>
 #include <preseg_passes/mark_aliases_prepare.h>
@@ -17,6 +16,7 @@
 #include <runtime/executor.h>
 #include <runtime/fusion_executor_cache.h>
 #include <scheduler/all_schedulers.h>
+#include <scheduler/tools/inlining.h>
 #include <scheduler/transpose.h>
 #include <scheduler/utils.h>
 #include <tests/cpp/utils.h>
