+++ tests/cpp/test_id_model.cpp
@@ -18,9 +18,9 @@
 #include <id_model/loop_promotion.h>
 #include <id_model/schedule.h>
 #include <id_model/to_string.h>
-#include <inlining.h>
 #include <ir/graphviz.h>
 #include <ops/all_ops.h>
+#include <scheduler/tools/inlining.h>
 #include <transform_iter.h>
 #include <val_graph_visitor.h>
