+++ tests/cpp/test_memory.cpp
@@ -13,7 +13,6 @@
 
 #include <debug.h>
 #include <fusion.h>
-#include <inlining.h>
 #include <ir/utils.h>
 #include <ops/alias.h>
 #include <ops/arith.h>
@@ -21,6 +20,7 @@
 #include <options.h>
 #include <scheduler/cache_policy_refiner.h>
 #include <scheduler/mma_utils.h>
+#include <scheduler/tools/inlining.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
 #include <type.h>
