+++ tests/cpp/test_scatter_gather.cpp
@@ -11,13 +11,13 @@
 #include <torch/torch.h>
 
 #include <exceptions.h>
-#include <inlining.h>
 #include <ir/all_nodes.h>
 #include <ir/builder.h>
 #include <ops/all_ops.h>
 #include <runtime/executor.h>
 #include <runtime/fusion_executor_cache.h>
 #include <scheduler/all_schedulers.h>
+#include <scheduler/tools/inlining.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
