+++ tests/cpp/test_tutorial.cpp
@@ -10,11 +10,11 @@
 
 #include <fusion.h>
 #include <id_model/id_model.h>
-#include <inlining.h>
 #include <ir/utils.h>
 #include <ops/alias.h>
 #include <ops/arith.h>
 #include <ops/utils.h>
+#include <scheduler/tools/inlining.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
 #include <type.h>
